+++ src/main/java/culturemedia/service/CulturotecaService.java
@@ -1,7 +1,6 @@
-package culturemedia.repository;
+package culturemedia.service;
 
 import culturemedia.exception.DurationNotValidException;
-import culturemedia.exception.VideoNotFoundException;
 import culturemedia.model.Video;
 import culturemedia.model.View;
